Add files via upload
The Previous version of the circuit has a problem when the hair resistance is smaller than the three reference resistors in the bridge, the output voltage goes negative, which is unusable by Adruino ADC directly. Thus a summer is added which shifts the output voltage by +2.5V. The output now has a voltage range of 0 to 5 V. Also, considering that simple clamp the hair together may not result in a good connection as the inner layers of the hair is not connected with the circuit directly, metal with low melting point (prevent damaging hair sample) should be used to fill in the gap between hair threads.