Add files via upload
This first iteration uses a Wheatstone bridge with the hair bundle in one leg to obtain an accurate measurement of voltage. Voltage followers built with high input impedance op-amps are used to "isolate" the rest of the circuit from the high resistance of the bridge. Finally, an amplifier setup is used to obtain the difference between two measurement points and outputs to an Arduino ADC. The resistors are selected such that output resistance is significantly smaller than Arduino ADC input impedance of 100 Mohms.